requirements.txt
Django>=4.2
djangorestframework>=3.14
uvicorn>=0.22.0
djangorestframework-simplejwt
drf-yasg


